requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.12
clang==5.0
cloudpickle==2.0.0
cycler==0.11.0
dataclasses==0.8
deepdiff==5.7.0
enum34==1.1.10
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.44.0
gym==0.21.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
mock==4.0.3
mpyq==0.2.5
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
ordered-set==4.0.2
pandas==1.1.5
Pillow==8.4.0
portpicker==1.5.0
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.8
PySC2==3.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s2clientprotocol==5.0.9.87702.0
s2protocol==5.0.9.87702.0
scipy==1.5.4
six==1.15.0
sk-video==1.1.10
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.9
websocket-client==1.3.1
Werkzeug==2.0.3
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
